requirements.txt
opencv-python
numpy
Pillow
pypdfium2
